move helper fxns to test suite
